+++ src/IO/WriteHelpers.h
@@ -139,7 +139,7 @@ inline void writeBoolText(bool x, WriteBuffer & buf)
 template <typename T>
 inline size_t writeFloatTextFastPath(T x, char * buffer)
 {
-    int result = 0;
+    Int64 result = 0;
 
     if constexpr (std::is_same_v<T, double>)
     {
