+++ src/Storages/HDFS/WriteBufferFromHDFS.cpp
@@ -5,6 +5,7 @@
 #include <Storages/HDFS/WriteBufferFromHDFS.h>
 #include <Storages/HDFS/HDFSCommon.h>
 #include <Common/Throttler.h>
+#include <Common/safe_cast.h>
 #include <hdfs/hdfs.h>
 
 namespace DB
@@ -57,7 +58,7 @@ struct WriteBufferFromHDFS::WriteBufferFromHDFSImpl
 
     int write(const char * start, size_t size) const
     {
-        int bytes_written = hdfsWrite(fs.get(), fout, start, size);
+        int bytes_written = hdfsWrite(fs.get(), fout, start, safe_cast<int>(size));
         if (write_settings.remote_throttler)
             write_settings.remote_throttler->add(bytes_written);
