+++ src/Storages/MergeTree/MergeTreeDataSelectExecutor.h
@@ -33,7 +33,7 @@ public:
         const SelectQueryInfo & query_info,
         ContextPtr context,
         UInt64 max_block_size,
-        unsigned num_streams,
+        size_t num_streams,
         QueryProcessingStage::Enum processed_stage,
         std::shared_ptr<PartitionIdToMaxBlock> max_block_numbers_to_read = nullptr,
         bool enable_parallel_reading = false) const;
@@ -46,7 +46,7 @@ public:
         const SelectQueryInfo & query_info,
         ContextPtr context,
         UInt64 max_block_size,
-        unsigned num_streams,
+        size_t num_streams,
         std::shared_ptr<PartitionIdToMaxBlock> max_block_numbers_to_read = nullptr,
         MergeTreeDataSelectAnalysisResultPtr merge_tree_select_result_ptr = nullptr,
         bool enable_parallel_reading = false) const;
@@ -62,7 +62,7 @@ public:
         const SelectQueryInfo & query_info,
         const ActionDAGNodes & added_filter_nodes,
         ContextPtr context,
-        unsigned num_streams,
+        size_t num_streams,
         std::shared_ptr<PartitionIdToMaxBlock> max_block_numbers_to_read = nullptr) const;
 
 private:
